make #next robust in face of redis network error
Current implementation is to rpop the requested
number of voters from the available list. When
redis client sends rpop command, it may fail
due to network error. Wrap each call to rpop in
begin/rescue and retry each rpop a max of 3 times
before moving on to the next load attempt. So for
each requested voter, we give redis client X
chances to return from rpop w/out a network error.

If 3 voters are requested and 3 network errors occur
each for the first 2 voters requested but no network
error (or less than X network errors) occur when
requesting the 3rd voter, that 3rd voter will be returned.
This works regardless of the order of errors. For
example, the 1st voter may load successfully,
the 2nd fail and the 3rd load successfully resulting
in 2 voters return (1st & 3rd), the 2nd voter
will be loaded the next time the cache needs populating.