stop throwing out phones-only answer data for unknown voters
When submitting a response, the data would be
pushed to redis only if a voter_id was present.
The voter_id is set when voter connects but
could feasibly be missing due to unforeseen app error
so now we also save the household_id (which is set
when dial is made) as a fallback and persist
submitted answers regardless of missing data.
When answer data is processed incomplete data
will be silo’d & kept in redis for future inspection.